+++ src/lib/graphql/download/DownloadFragments.ts
@@ -1,25 +1,32 @@
 import gql from 'graphql-tag';
-import { CHAPTER_BASE_FIELDS } from '@/lib/graphql/chapter/ChapterFragments.ts';
+import { SOURCE_BASE_FIELDS } from '@/lib/graphql/source/SourceFragments.ts';
 import { MANGA_BASE_FIELDS } from '@/lib/graphql/manga/MangaFragments.ts';
+import { CHAPTER_BASE_FIELDS } from '@/lib/graphql/chapter/ChapterFragments.ts';
 
 export const DOWNLOAD_TYPE_FIELDS = gql`
-    ${CHAPTER_BASE_FIELDS}
     ${MANGA_BASE_FIELDS}
+    ${SOURCE_BASE_FIELDS}
+    ${CHAPTER_BASE_FIELDS}
 
     fragment DOWNLOAD_TYPE_FIELDS on DownloadType {
         chapter {
             ...CHAPTER_BASE_FIELDS
             isDownloaded
+            pageCount
         }
 
         manga {
             ...MANGA_BASE_FIELDS
             downloadCount
+            source {
+                ...SOURCE_BASE_FIELDS
+            }
         }
 
         progress
         state
         tries
+        position
     }
 `;
 
@@ -49,7 +56,6 @@ export const DOWNLOAD_UPDATES_FIELDS = gql`
             type
             download {
                 ...DOWNLOAD_TYPE_FIELDS
-                position
             }
         }
     }