+++ src/lib/graphql/server/ServerInfoQuery.ts
@@ -10,6 +10,21 @@ export const GET_ABOUT = gql`
             repoUrl
             name
             version
+            platformInfo {
+                os {
+                    name
+                    build
+                    version
+                }
+                arch
+                headless
+                jvm {
+                    javaVersion
+                    vmName
+                    vmVendor
+                    vmVersion
+                }
+            }
         }
         aboutWebUI {
             ...ABOUT_WEBUI